connected att to cart functionality to catalog and cart items count
