+++ index.html
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="./dist/styles.min.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
     <script src="./src/js/app.js" type="module"></script>
-    <script src="./src/js/addToCartFunc.js" type="module"></script>
+    <script src="./src/js/cartFunctions.js" type="module"></script>
 
     <!-- Google tag (gtag.js) -->
     <script async src="https://www.googletagmanager.com/gtag/js?id=G-LCST1VL486"></script>
@@ -38,7 +38,7 @@
           <a href="./pages/cart.html" class="cta-inverted" id="cart-button">
             Cart
             <img src="./src/images/icons/shopping-bag-drk.svg" width="24" height="24" alt="Cart">
-            <span class="count-badge hide" id="item-count">1</span>
+            <span class="count-badge " id="item-count"></span>
           </a>
         </div>
       </div>
