+++ pages/cart.html
@@ -47,6 +47,6 @@
    </main>
 
    <script src="/src/js/cart.js" type="module"></script>
-    <script src="/src/js/addToCartFunc.js" type="module"></script>
+    <script src="/src/js/cartFunctions.js" type="module"></script>
 </body>
 </html>
