+++ src/js/app.js
@@ -1,9 +1,15 @@
 import { getProducts } from "./services/apiService.js";
+import { getById, cartItemCountLS} from "./cartFunctions.js";
 
 const filterSelectEl = document.getElementById("filter-select");
 const sortSelectEl = document.getElementById("sort-select");
 const itemsContainerEl = document.getElementById("items-container");
 const loaderEl = document.getElementById("loader");
+let cartCount = document.querySelector("#item-count") 
+
+let showItemsCount = () => {  // count the items in cart and show the count on the cart 
+  cartCount.innerHTML = cartItemCountLS()
+}
 
 let fetchedProducts = null;
 let unsortedProducts = null;
@@ -16,6 +22,7 @@ const loadProducts = async () => {
     fetchedProducts = await getProducts();
     unsortedProducts = Array.from(fetchedProducts);
     displayProducts(fetchedProducts);
+    showItemsCount()
   } catch (error) {
     console.log(error);
     itemsContainerEl.innerHTML = `<p>Failed loading products. <br> Try Agin later.</p>`;
@@ -49,7 +56,7 @@ const displayProducts = () => {
     )
     .map(
       (item) =>
-        `<article class="product">
+        `<article class="product" data-key="${item.index}" >
           <img class="product-img" src="${item.image}" alt="${item.title}" width="150" height="175" />
           <h3 class="truncate">${item.title}</h3>
           <p class="truncate">${item.description}</p>
@@ -67,11 +74,16 @@ const displayProducts = () => {
   itemsContainerEl.classList.remove("hide");
   loaderEl.classList.add("hide");
 
+
+
   const productsButtons = itemsContainerEl.querySelectorAll("button");
   productsButtons.forEach((button) => {
-    button.addEventListener("click", () => {
+    button.addEventListener("click", (event) => {
+      const product = parseInt(event.target.getAttribute("data-id"))
       // Catch Product ID
       console.log(button.dataset.id);
+      getById(fetchedProducts, product)
+
     })
   });
 }
